+++ src/templatemanager/base/actionsModal/actionsModal.tsx
@@ -1,0 +1,8 @@
+import './actionsModal.scss';
+import { h } from 'preact';
+
+export function ActionsBottomSlider(props: any) {
+    return (
+        <div>Modal</div>
+    );
+}