+++ src/templatemanager/templates/v3/table/table.tsx
@@ -43,74 +43,34 @@ export function Table(props: any) {
     }
 
     if (msgData.message?.[0]?.component?.payload?.template_type == 'table' && msgData.message[0].component.payload?.table_design == 'responsive') {
+        const selectItem = (e: any) => {
+            e.currentTarget.classList.toggle('acc-open');
+        }
         return (
             <Fragment>
                 <div className='table-container'>
                     <Message {...messageobj} />
-                    {/* { msgData.message[0].component.payload.elements.map((ele: any) => (
-    <div style={{display: 'flex'}}>
-        { ele.Values.map((e: any, i: any) => (
-            <div className={i >= 2 ? 'hide': ''}>
-                <div className={i < 2 ? 'hide' : ''}>{msgData.message[0].component.payload.columns[i]}</div>
-                <div>{e}</div>
-            </div>
-        ))}
-    </div>
-))} */}
                     <div className="table-response-wrapper-container">
                         <div className="table-response-wrapper">
-                            <div className="acc-block-content">
-                                <div className="info-blcok">
-                                    <h1>SI</h1>
-                                    <p>1</p>
+                        { msgData.message[0].component.payload.elements.map((ele: any) => (
+                            <div className="acc-block-content" onClick={selectItem}>
+                                { ele.Values.map((e: any, i: any) => ( i < 2 &&<div className="info-block">
+                                    <h1 className="hide">{msgData.message[0].component.payload.columns[i]}</h1>
+                                    <p>{e}</p>
                                 </div>
-                                <div className="info-blcok">
-                                    <h1>Name</h1>
-                                    <p>Peter</p>
+                                ))}
+
+                                { ele.Values.map((e: any, i: any) => ( i >= 2 &&<div className="info-block hide">
+                                    <h1>{msgData.message[0].component.payload.columns[i]}</h1>
+                                    <p>{e}</p>
                                 </div>
+                                ))}
+
                                 <div className="icon-block">
                                     <i className="sdkv3-cheveron-right"></i>
                                 </div>
                             </div>
-                            <div className="acc-block-content">
-                                <div className="info-blcok">
-                                    <h1>SI</h1>
-                                    <p>1</p>
-                                </div>
-                                <div className="info-blcok">
-                                    <h1>Name</h1>
-                                    <p>Peter</p>
-                                </div>
-                                <div className="icon-block">
-                                    <i className="sdkv3-cheveron-right"></i>
-                                </div>
-                            </div>
-                            <div className="acc-block-content">
-                                <div className="info-blcok">
-                                    <h1>SI</h1>
-                                    <p>1</p>
-                                </div>
-                                <div className="info-blcok">
-                                    <h1>Name</h1>
-                                    <p>Peter</p>
-                                </div>
-                                <div className="icon-block">
-                                    <i className="sdkv3-cheveron-right"></i>
-                                </div>
-                            </div>
-                            <div className="acc-block-content">
-                                <div className="info-blcok">
-                                    <h1>SI</h1>
-                                    <p>1</p>
-                                </div>
-                                <div className="info-blcok">
-                                    <h1>Name</h1>
-                                    <p>Peter</p>
-                                </div>
-                                <div className="icon-block">
-                                    <i className="sdkv3-cheveron-right"></i>
-                                </div>
-                            </div>
+                        ))}
                         </div>
                         <button className="show-more-btn">Show More</button>
                     </div>