+++ CHANGELOG.md
@@ -1,4 +1,20 @@
+### 10.1.9
 
+- SDK Release Aug 19, 2023
+* Bug fixes
+  * carousel template bugfixes with multiple 
+  * quick reply template bugfixes
+  * Bugfix for back to back form trigger within 1 sec with continuous loading
+  * Added a config to send acknowledgement on bot response.Configuration available at chatConfig.botOptions.enableAck
+### 10.1.1
+
+- SDK Release July 4, 2023
+* Bug fixes
+  * Reconnect on token expiry 
+  * ADA mac safari voiceover bug fixes
+  * Support for .gif fileType
+  * Multiple carouselTemplate fix
+  * Solutions templates plugin added
 ### 10.1.0
 
 - Release Apr 16, 2023